+++ python_modules/dagster-graphql/dagster_graphql/schema/roots/subscription.py
@@ -7,12 +7,11 @@ from dagster import AssetKey
 from dagster._core.storage.compute_log_manager import ComputeIOType
 
 from ...implementation.execution import get_compute_log_observable, get_pipeline_run_observable
-from ..external import (
+from ..asset_subscription import (
     GrapheneAssetLogEventsSubscriptionPayload,
-    GrapheneLocationStateChangeSubscription,
     get_asset_log_events_observable,
-    get_location_state_change_observable,
 )
+from ..external import GrapheneLocationStateChangeSubscription, get_location_state_change_observable
 from ..logs.compute_logs import GrapheneComputeIOType, GrapheneComputeLogFile
 from ..pipelines.subscription import GraphenePipelineRunLogsSubscriptionPayload
